Split create issue modal content and modal
Fixes a bug where the project id was not prefilled
